# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "APIphoto_server",
        "contact": {},
        "version": "V1.0"
    },
    "paths": {
        "/api/photo/delete": {
            "post": {
                "description": "删除用户上传的图片",
                "tags": [
                    "Photo control"
                ],
                "summary": "删除用户上传的图片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "图片名称",
                        "name": "PhotoName",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/photo/upload": {
            "post": {
                "description": "接收用户上传的图片文件并存储到指定用户目录，同时记录文件元信息到数据库",
                "tags": [
                    "Photo control"
                ],
                "summary": "用户上传图片接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "要上传的图片文件(支持JPG/PNG)",
                        "name": "photo",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "图片描述信息",
                        "name": "description",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/CreateUser": {
            "post": {
                "tags": [
                    "user management"
                ],
                "summary": "add user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserName",
                        "name": "UserName",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "Password",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "RePassword",
                        "name": "RePassword",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/LoginInUser": {
            "post": {
                "tags": [
                    "user management"
                ],
                "summary": "login in user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserName",
                        "name": "UserName",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "Password",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\",\"id\",\"token\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "code\",\"message\",\"id\",\"token\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/download_user-basic-message": {
            "post": {
                "description": "获取用户基本信息（包括用户名、手机号和邮箱）",
                "tags": [
                    "user management"
                ],
                "summary": "获取用户基本信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user/download_user-homepage-message": {
            "post": {
                "description": "获取当前登录用户的主页信息（包括用户名、头像和自我介绍）",
                "tags": [
                    "User Home Message"
                ],
                "summary": "获取用户主页信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user/upload-avatar": {
            "post": {
                "description": "上传用户头像接口（支持JPEG/PNG格式，最大30MB）",
                "tags": [
                    "User Home Message"
                ],
                "summary": "上传用户头像",
                "parameters": [
                    {
                        "type": "file",
                        "description": "头像文件（支持JPEG/PNG格式）",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "413": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/upload_self-introduce": {
            "post": {
                "description": "用户更新自我介绍信息接口（需Token认证）",
                "tags": [
                    "User Home Message"
                ],
                "summary": "更新用户自我介绍",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户自我介绍内容",
                        "name": "selfIntroduce",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\", \"id\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "code\", \"id\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\", \"id\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/upload_user-email": {
            "post": {
                "description": "用户更新用户邮箱接口（需Token认证）",
                "tags": [
                    "user management"
                ],
                "summary": "更新用户邮箱",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\", \"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "code\", \"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\", \"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/upload_user-phone": {
            "post": {
                "description": "用户更新用户电话号接口（需Token认证）",
                "tags": [
                    "user management"
                ],
                "summary": "更新用户电话号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户电话号",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\", \"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "code\", \"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\", \"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}